DEV-53 fix bug test learning
1. LSF create_at can happen before card create in testing environment...
making unreasonable comparison when querying cards to learn
So, set learn_date of LSF to 23:59:59
2. test_learning should mimic flow of user, which comprises of...
navigating through before_learning
There is still a bug about the existence of LSF object state after...
sending a GET request to /before_learning (which should not because...
the route does not write new LSF into DB).
This bizarre happens only in testing env for now.